Fix a bunch of packet structures

+++ WowPacketParser/Parsing/Parsers/NpcHandler.cs
@@ -576,7 +576,7 @@ namespace WowPacketParser.Parsing.Parsers
                     OptionIndex = gossipOption.OptionID.Value,
                     OptionNpc = (int)gossipOption.OptionNpc,
                     BoxCoded = gossipOption.BoxCoded.Value,
-                    BoxCost = gossipOption.BoxMoney.Value,
+                    BoxCost = (uint)gossipOption.BoxMoney.Value,
                     Text = gossipOption.OptionText,
                     BoxText = gossipOption.BoxText
                 });
