+++ WowPacketParserModule.V5_4_0_17359/Parsers/NpcHandler.cs
@@ -151,7 +151,7 @@ namespace WowPacketParserModule.V5_4_0_17359.Parsers
                     OptionIndex = gossipOption.OptionID.Value,
                     OptionNpc = (int)gossipOption.OptionNpc,
                     BoxCoded = gossipOption.BoxCoded.Value,
-                    BoxCost = gossipOption.BoxMoney.Value,
+                    BoxCost = (uint)gossipOption.BoxMoney.Value,
                     Text = gossipOption.OptionText,
                     BoxText = gossipOption.BoxText
                 });
