+++ WowPacketParserModule.V5_4_1_17538/Parsers/NpcHandler.cs
@@ -88,7 +88,7 @@ namespace WowPacketParserModule.V5_4_1_17538.Parsers
                     OptionIndex = gossipOption.OptionID.Value,
                     OptionNpc = (int)gossipOption.OptionNpc,
                     BoxCoded = gossipOption.BoxCoded.Value,
-                    BoxCost = gossipOption.BoxMoney.Value,
+                    BoxCost = (uint)gossipOption.BoxMoney.Value,
                     Text = gossipOption.OptionText,
                     BoxText = gossipOption.BoxText
                 });
